Pull request 117: AG-37593, Understand User filter (& DNS) ID
Merge in ADGUARD-CORE-LIBS/filter-list-manager from feature/AG-37593 to master

Squashed commit of the following:

commit b689a925434fcc477d4a34ef675f1c0010096004
Author: Mikhail Berdnikov <m.berdnikov@adguard.com>
Date:   Thu Nov 21 02:20:35 2024 +0200

    AG-37593, base

+++ platform/windows/AdGuard.FilterListManager/MarshalLogic/ActiveRulesInfo.cs
@@ -10,7 +10,7 @@ namespace AdGuard.FilterListManager.MarshalLogic
         /// <summary>
         /// Gets the filter identifier.
         /// </summary>
-        public long FilterId { get; }
+        public int FilterId { get; }
 
         /// <summary>
         /// Gets the group identifier.
@@ -37,7 +37,7 @@ namespace AdGuard.FilterListManager.MarshalLogic
         /// <param name="groupId">The group identifier.</param>
         /// <param name="isTrusted">if set to <c>true</c> this filter is trusted.</param>
         /// <param name="rules">The rules of the filter.</param>
-        public ActiveRulesInfo(long filterId, int groupId, bool isTrusted, List<string> rules)
+        public ActiveRulesInfo(int filterId, int groupId, bool isTrusted, List<string> rules)
         {
             FilterId = filterId;
             GroupId = groupId;
