+++ platform/windows/AdGuard.FilterListManager/MarshalLogic/UpdateFilterError.cs
@@ -10,7 +10,7 @@ namespace AdGuard.FilterListManager.MarshalLogic
         /// </summary>
         /// <param name="filterId">The filter identifier.</param>
         /// <param name="message">The message of the error.</param>
-        public UpdateFilterError(long filterId, string message)
+        public UpdateFilterError(int filterId, string message)
         {
             this.filterId = filterId;
             this.message = message;
@@ -19,7 +19,7 @@ namespace AdGuard.FilterListManager.MarshalLogic
         /// <summary>
         /// Gets or sets the filter identifier.
         /// </summary>
-        public long filterId { get; set; }
+        public int filterId { get; set; }
 
         /// <summary>
         /// Gets or sets the message of the error.
